Refactor packages for easier import
