+++ cmd/nitro/nitro.go
@@ -7,6 +7,7 @@ import (
 	"context"
 	"crypto/ecdsa"
 	"fmt"
+	"github.com/offchainlabs/nitro/cmd/genericconf"
 	"math"
 	"math/big"
 	"os"
@@ -62,7 +63,7 @@ func main() {
 
 		return
 	}
-	logFormat, err := conf.ParseLogType(nodeConfig.LogType)
+	logFormat, err := genericconf.ParseLogType(nodeConfig.LogType)
 	if err != nil {
 		flag.Usage()
 		panic(fmt.Sprintf("Error parsing log type: %v", err))
@@ -334,53 +335,53 @@ func main() {
 }
 
 type NodeConfig struct {
-	Conf          conf.ConfConfig          `koanf:"conf"`
-	Node          arbnode.Config           `koanf:"node"`
-	L1            conf.L1Config            `koanf:"l1"`
-	L2            conf.L2Config            `koanf:"l2"`
-	LogLevel      int                      `koanf:"log-level"`
-	LogType       string                   `koanf:"log-type"`
-	Persistent    conf.PersistentConfig    `koanf:"persistent"`
-	HTTP          conf.HTTPConfig          `koanf:"http"`
-	WS            conf.WSConfig            `koanf:"ws"`
-	DevInit       bool                     `koanf:"dev-init"`
-	NoInit        bool                     `koanf:"no-init"`
-	ImportFile    string                   `koanf:"import-file"`
-	Metrics       bool                     `koanf:"metrics"`
-	MetricsServer conf.MetricsServerConfig `koanf:"metrics-server"`
+	Conf          genericconf.ConfConfig          `koanf:"conf"`
+	Node          arbnode.Config                  `koanf:"node"`
+	L1            conf.L1Config                   `koanf:"l1"`
+	L2            conf.L2Config                   `koanf:"l2"`
+	LogLevel      int                             `koanf:"log-level"`
+	LogType       string                          `koanf:"log-type"`
+	Persistent    conf.PersistentConfig           `koanf:"persistent"`
+	HTTP          genericconf.HTTPConfig          `koanf:"http"`
+	WS            genericconf.WSConfig            `koanf:"ws"`
+	DevInit       bool                            `koanf:"dev-init"`
+	NoInit        bool                            `koanf:"no-init"`
+	ImportFile    string                          `koanf:"import-file"`
+	Metrics       bool                            `koanf:"metrics"`
+	MetricsServer genericconf.MetricsServerConfig `koanf:"metrics-server"`
 }
 
 var NodeConfigDefault = NodeConfig{
-	Conf:          conf.ConfConfigDefault,
+	Conf:          genericconf.ConfConfigDefault,
 	Node:          arbnode.ConfigDefault,
 	L1:            conf.L1ConfigDefault,
 	L2:            conf.L2ConfigDefault,
 	LogLevel:      int(log.LvlInfo),
 	LogType:       "plaintext",
 	Persistent:    conf.PersistentConfigDefault,
-	HTTP:          conf.HTTPConfigDefault,
-	WS:            conf.WSConfigDefault,
+	HTTP:          genericconf.HTTPConfigDefault,
+	WS:            genericconf.WSConfigDefault,
 	DevInit:       false,
 	ImportFile:    "",
 	Metrics:       false,
-	MetricsServer: conf.MetricsServerConfigDefault,
+	MetricsServer: genericconf.MetricsServerConfigDefault,
 }
 
 func NodeConfigAddOptions(f *flag.FlagSet) {
-	conf.ConfConfigAddOptions("conf", f)
+	genericconf.ConfConfigAddOptions("conf", f)
 	arbnode.ConfigAddOptions("node", f, true, true)
 	conf.L1ConfigAddOptions("l1", f)
 	conf.L2ConfigAddOptions("l2", f)
 	f.Int("log-level", NodeConfigDefault.LogLevel, "log level")
 	f.String("log-type", NodeConfigDefault.LogType, "log type (plaintext or json)")
 	conf.PersistentConfigAddOptions("persistent", f)
-	conf.HTTPConfigAddOptions("http", f)
-	conf.WSConfigAddOptions("ws", f)
+	genericconf.HTTPConfigAddOptions("http", f)
+	genericconf.WSConfigAddOptions("ws", f)
 	f.Bool("dev-init", NodeConfigDefault.DevInit, "init with dev data (1 account with balance) instead of file import")
 	f.Bool("no-init", NodeConfigDefault.DevInit, "Do not init chain. Data must be valid in database.")
 	f.String("import-file", NodeConfigDefault.ImportFile, "path for json data to import")
 	f.Bool("metrics", NodeConfigDefault.Metrics, "enable metrics")
-	conf.MetricsServerAddOptions("metrics-server", f)
+	genericconf.MetricsServerAddOptions("metrics-server", f)
 }
 
 func (c *NodeConfig) ResolveDirectoryNames() error {
@@ -394,7 +395,7 @@ func (c *NodeConfig) ResolveDirectoryNames() error {
 	return nil
 }
 
-func ParseNode(ctx context.Context, args []string) (*NodeConfig, *conf.WalletConfig, *conf.WalletConfig, *ethclient.Client, *big.Int, error) {
+func ParseNode(ctx context.Context, args []string) (*NodeConfig, *genericconf.WalletConfig, *genericconf.WalletConfig, *ethclient.Client, *big.Int, error) {
 	f := flag.NewFlagSet("", flag.ContinueOnError)
 
 	NodeConfigAddOptions(f)
@@ -521,8 +522,8 @@ func ParseNode(ctx context.Context, args []string) (*NodeConfig, *conf.WalletCon
 	// Don't pass around wallet contents with normal configuration
 	l1Wallet := nodeConfig.L1.Wallet
 	l2DevWallet := nodeConfig.L2.DevWallet
-	nodeConfig.L1.Wallet = conf.WalletConfigDefault
-	nodeConfig.L2.DevWallet = conf.WalletConfigDefault
+	nodeConfig.L1.Wallet = genericconf.WalletConfigDefault
+	nodeConfig.L2.DevWallet = genericconf.WalletConfigDefault
 
 	return &nodeConfig, &l1Wallet, &l2DevWallet, l1Client, l1ChainId, nil
 }
