+++ cmd/relay/relay.go
@@ -6,6 +6,7 @@ package main
 import (
 	"context"
 	"fmt"
+	"github.com/offchainlabs/nitro/cmd/genericconf"
 	"net/http"
 	"os"
 	"os/signal"
@@ -15,7 +16,6 @@ import (
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/knadh/koanf/parsers/json"
 	"github.com/knadh/koanf/providers/confmap"
-	"github.com/offchainlabs/nitro/cmd/conf"
 	"github.com/offchainlabs/nitro/cmd/util"
 	"github.com/pkg/errors"
 	flag "github.com/spf13/pflag"
@@ -52,7 +52,7 @@ func startup() error {
 		return nil
 	}
 
-	logFormat, err := conf.ParseLogType(relayConfig.LogType)
+	logFormat, err := genericconf.ParseLogType(relayConfig.LogType)
 	if err != nil {
 		flag.Usage()
 		panic(fmt.Sprintf("Error parsing log type: %v", err))
@@ -93,21 +93,21 @@ func startup() error {
 }
 
 type RelayConfig struct {
-	Conf     conf.ConfConfig `koanf:"conf"`
-	LogLevel int             `koanf:"log-level"`
-	LogType  string          `koanf:"log-type"`
-	Node     RelayNodeConfig `koanf:"node"`
+	Conf     genericconf.ConfConfig `koanf:"conf"`
+	LogLevel int                    `koanf:"log-level"`
+	LogType  string                 `koanf:"log-type"`
+	Node     RelayNodeConfig        `koanf:"node"`
 }
 
 var RelayConfigDefault = RelayConfig{
-	Conf:     conf.ConfConfigDefault,
+	Conf:     genericconf.ConfConfigDefault,
 	LogLevel: int(log.LvlInfo),
 	LogType:  "plaintext",
 	Node:     RelayNodeConfigDefault,
 }
 
 func RelayConfigAddOptions(f *flag.FlagSet) {
-	conf.ConfConfigAddOptions("conf", f)
+	genericconf.ConfConfigAddOptions("conf", f)
 	f.Int("log-level", RelayConfigDefault.LogLevel, "log level")
 	f.String("log-type", RelayConfigDefault.LogType, "log type")
 	RelayNodeConfigAddOptions("node", f)
